refactor to use get_clean_diff instead of get_diff
Now, the output will be cleaned by default. We can also change this
later, if the need for metadata becomes apparent or the LLM requires it
for some reason.

To be honest, I'm very proud the codebase is this simple and I can
very easily refactor and enhance without really banging my head on a
wall.