add .gitattributes to normalize commited changes
We also use line-ending = "lf" in configuring ruff's formatter. This is
to make sure even if they didn't run ruff format . before commiting,
they won't mess up. There's also the packaging functionality in which we
make it so that the zipped result of the repo doesn't include junk
metadata like .gitignore or .gitattributes itself.